+++ src/util/identifier.cpp
@@ -6,7 +6,7 @@ Author: Daniel Kroening, kroening@kroening.com
 
 \*******************************************************************/
 
-#include <string.h>
+#include <cstring>
 
 #include "identifier.h"
 
@@ -52,7 +52,7 @@ void identifiert::parse(const std::string &s)
 {
   std::string component;
 
-  for(unsigned i=0; i<s.size();)
+  for(size_t i=0; i<s.size();)
   {
     for(; i<s.size(); i++)
     {
